Improves keymap with tap-hold and gaming layer
Enhances the keymap by introducing a tap-hold Ctrl/Win key and refining the gaming layer.

The Ctrl/Win key allows for quick Ctrl actions on tap and GUI/Meta actions on hold.
The gaming layer is updated for better key placement and functionality.
Tapping Term is set to per key